Revert "Merge branch 'feature/parallel-txs' into bal-devnet-3"
This reverts commit 8e8d690d3d75cfe423990f9fd66eb8e9e0a56500, reversing
changes made to 0b0c249d2600614f4623920bb184b612317b3f40.

+++ src/Nethermind/Nethermind.Core.Test/Encoding/BlockAccessListDecoderTests.cs
@@ -63,7 +63,7 @@ public class BlockAccessListDecoderTests
     public void Can_decode_then_encode_slot_change()
     {
         StorageChange parentHashStorageChange = new(0, new UInt256(Bytes.FromHexString("0xc382836f81d7e4055a0e280268371e17cc69a531efe2abee082e9b922d6050fd"), isBigEndian: true));
-        SlotChanges expected = new(0, new SortedList<int, StorageChange> { { 0, parentHashStorageChange } });
+        SlotChanges expected = new(0, new SortedList<ushort, StorageChange> { { 0, parentHashStorageChange } });
 
         string expectedRlp = "0x" + Bytes.ToHexString(Rlp.Encode(expected).Bytes);
 
@@ -136,7 +136,7 @@ public class BlockAccessListDecoderTests
         StorageChange storageChangeDecoded = Rlp.Decode<StorageChange>(storageChangeBytes, RlpBehaviors.None);
         Assert.That(storageChange, Is.EqualTo(storageChangeDecoded));
 
-        var storageChanges = new SortedList<int, StorageChange> { { 10, storageChange } };
+        var storageChanges = new SortedList<ushort, StorageChange> { { 10, storageChange } };
         SlotChanges slotChanges = new(0xbad, storageChanges);
         byte[] slotChangesBytes = Rlp.Encode(slotChanges, RlpBehaviors.None).Bytes;
         SlotChanges slotChangesDecoded = Rlp.Decode<SlotChanges>(slotChangesBytes, RlpBehaviors.None);
@@ -258,7 +258,7 @@ public class BlockAccessListDecoderTests
     [Test]
     public void Decoding_account_changes_with_unsorted_balance_changes_throws()
     {
-        SortedList<int, BalanceChange> balanceChanges = new(DescendingComparer<int>())
+        SortedList<ushort, BalanceChange> balanceChanges = new(DescendingComparer<ushort>())
         {
             { 1, new(1, UInt256.One) },
             { 2, new(2, UInt256.Zero) }
@@ -281,7 +281,7 @@ public class BlockAccessListDecoderTests
     [Test]
     public void Decoding_account_changes_with_unsorted_nonce_changes_throws()
     {
-        SortedList<int, NonceChange> nonceChanges = new(DescendingComparer<int>())
+        SortedList<ushort, NonceChange> nonceChanges = new(DescendingComparer<ushort>())
         {
             { 1, new(1, 1) },
             { 2, new(2, 2) }
@@ -304,7 +304,7 @@ public class BlockAccessListDecoderTests
     [Test]
     public void Decoding_account_changes_with_unsorted_code_changes_throws()
     {
-        SortedList<int, CodeChange> codeChanges = new(DescendingComparer<int>())
+        SortedList<ushort, CodeChange> codeChanges = new(DescendingComparer<ushort>())
         {
             { 1, new(1, [0x01]) },
             { 2, new(2, [0x02]) }
@@ -327,7 +327,7 @@ public class BlockAccessListDecoderTests
     [Test]
     public void Decoding_slot_changes_with_unsorted_storage_changes_throws()
     {
-        SortedList<int, StorageChange> storageChanges = new(DescendingComparer<int>())
+        SortedList<ushort, StorageChange> storageChanges = new(DescendingComparer<ushort>())
         {
             { 1, new(1, UInt256.One) },
             { 2, new(2, UInt256.Zero) }
